requirements.txt
polars
pandas
sqlalchemy
python-dotenv
mysql-connector-python
catboost
scikit-learn
networkx
plotly
